Moved ImageUrlWithData to appropriate package
